I got stuff to compile...
However, I ripped so much of my old "scene" management out that I can no
longer run anything. Will need to integrate the rendering code so I can
at least run the program.

Also, I made sure there were 0 warnings when doing a full build.
(silencing the 3rd party stuff was hard.)

+++ Headers/Sphere.hpp
@@ -3,7 +3,7 @@
 
 #include "Object.hpp"
 
-const GLfloat PI = 4.0*atan(1.0);
+const GLdouble PI = 4.0*atan(1.0);
 
 class Sphere : public Object {
 public:
